% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_colors.R
\name{bayes_colors}
\alias{bayes_colors}
\title{Function to retrieve bayesEO color table}
\usage{
bayes_colors()
}
\value{
A tibble with color names and values
}
\description{
Returns a color table
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
