% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_colors.R
\name{bayes_colors_show}
\alias{bayes_colors_show}
\title{Function to show colors in SITS}
\usage{
bayes_colors_show()
}
\value{
no return, called for side effects
}
\description{
Shows the default SITS colors
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
