% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_plot.R
\name{bayes_plot}
\alias{bayes_plot}
\title{Plot probability and variance maps}
\usage{
bayes_plot(
  x,
  scale = 1,
  labels = NULL,
  quantile = NULL,
  palette = "YlGnBu",
  tmap_legend_title_size = 1,
  tmap_legend_text_size = 1,
  tmap_legend_bg_color = "white",
  tmap_legend_bg_alpha = 0.5
)
}
\arguments{
\item{x}{SpatRaster to be plotted.}

\item{scale}{Scaling factor to apply to the data}

\item{labels}{Labels to be plotted}

\item{quantile}{Thereshold of values to be plotted}

\item{palette}{An RColorBrewer palette}

\item{tmap_legend_title_size}{Size of legend title (default: 1.5)}

\item{tmap_legend_text_size}{Size of legend text (default: 1.2)}

\item{tmap_legend_bg_color}{Color of legend backgound (default: "white")}

\item{tmap_legend_bg_alpha}{Transparency of legend background (default: 0.5)}
}
\value{
A plot object
}
\description{
Plot probability and variance maps
}
\examples{
if (bayes_run_examples()) {
    # get the probability file
    data_dir <- system.file("/extdata/probs/", package = "bayesEO")
    file <- list.files(data_dir)
    # read the probability file into a SpatRaster
    x <- terra::rast(paste0(data_dir, "/", file))
    # include the labels
    labels <- c("Water", "ClearCut_Burn", "ClearCut_Soil",
             "ClearCut_Veg", "Forest", "Wetland")
    # associate the labels to the names of the SpatRaster
    names(x) <- labels
    # Plot the probability image
    bayes_plot(x, scale = 0.0001, labels = c("Forest", "ClearCut_Soil"))
}

}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
