% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_read.R
\name{bayes_read_image}
\alias{bayes_read_image}
\title{Read probability maps}
\usage{
bayes_read_image(red_file, green_file, blue_file)
}
\arguments{
\item{file}{Full path to raster multi-band file
containing probability matrices}

\item{labels}{Labels to be assigned to the bands}
}
\value{
A SpatRaster object
}
\description{
Read probability maps
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
