% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_utils.R
\name{bayes_run_examples}
\alias{bayes_run_examples}
\title{Informs if  examples should run}
\usage{
bayes_run_examples()
}
\value{
A logical value
}
\description{
This function informs if examples should run.
To run the examples, set "BAYES_RUN_EXAMPLES" environment
variable to "YES" using
Sys.setenv("BAYES_RUN_EXAMPLES" = "YES")
To come back to the default behaviour, please unset
the enviroment variable
Sys.unsetenv("BAYES_RUN_EXAMPLES")
}
