% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_utils.R
\name{bayes_run_tests}
\alias{bayes_run_tests}
\title{Informs if tests should run}
\usage{
bayes_run_tests()
}
\value{
TRUE/FALSE
}
\description{
This function informs if tests should run.
To run the examples, set "BAYES_RUN_TESTS" environment
variable to "YES" using
Sys.setenv("BAYES_RUN_TESTS" = "YES")
To come back to the default behaviour, please unset
the enviroment variable
Sys.unsetenv("BAYES_RUN_TESTS")
}
