% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_plot.R
\name{bayes_plot_rgb}
\alias{bayes_plot_rgb}
\title{Plot RGB data cubes}
\usage{
bayes_plot_rgb(image, red, green, blue)
}
\arguments{
\item{image}{Object of class SpatRaster.}

\item{red}{Band for red color.}

\item{green}{Band for green color.}

\item{blue}{Band for blue color.}
}
\value{
A plot object with an RGB image
}
\description{
Plot RGB raster cube
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
