% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_view.R
\name{bayes_view}
\alias{bayes_view}
\title{Include leaflet to view images (BW or RGB)}
\usage{
bayes_view(
  image,
  map,
  red,
  green,
  blue,
  legend = NULL,
  palette = "Spectral",
  opacity = 0.5
)
}
\arguments{
\item{image}{Image to be plotted}

\item{map}{Classified map to be overlayed on top on image}

\item{red}{Band for red color.}

\item{green}{Band for green color.}

\item{blue}{Band for blue color.}

\item{legend}{Named vector that associates labels to colors.}

\item{palette}{Palette provided in the configuration file.}

\item{opacity}{Opacity of overlayed map}
}
\value{
A leaflet object.
}
\description{
Include leaflet to view images (BW or RGB)
}
\author{
Gilberto Camara, \email{gilberto.camara@inpe.br}
}
